catalog: split alpha entry point into multiple modules
